moving stuff around in archimedes
