+++ Gemfile
@@ -14,7 +14,7 @@ gem 'sdoc', '~> 0.4.0', group: :doc
 # Use ActiveModel has_secure_password
 # gem 'bcrypt', '~> 3.1.7'
 
-gem 'haml-rails', '>= 0.3.4'
+gem 'haml-rails', '>= 1.0.0'
 
 gem 'rltk'
 
@@ -35,6 +35,8 @@ group :development, :test do
 
   # Spring speeds up development by keeping your application running in the background. Read more: https://github.com/rails/spring
   gem 'spring'
+
+  gem 'rspec'
 end
 
 #gem 'devise'
