+++ app/models/archimedes.rb
@@ -1,0 +1,9 @@
+require_relative 'archimedes/manifest'
+require_relative 'archimedes/macro'
+require_relative 'archimedes/function'
+require_relative 'archimedes/infix_lexer'
+require_relative 'archimedes/infix_parser'
+require_relative 'archimedes/language_error'
+
+require_relative 'archimedes/matrix'
+require_relative 'archimedes/vector'
