+++ app/models/archimedes/manifest.rb
@@ -53,36 +53,55 @@
 #   }
 # }
 
-class DataManifest
-  attr_reader :vars
-
-  def initialize(token, project_name, manifest)
-    @manifest = manifest
-    @vars = {}
-    @env = InfixParser::Environment.create(token, project_name, @vars)
-  end
-
-  def fill
-    @manifest.each do |variable, query|
-      begin
-        @vars[variable] = resolve(query)
-      rescue RLTK::NotInLanguage => e
-        raise LanguageError, "Could not resolve @#{variable} = #{query}"
+module Archimedes
+  class Manifest < RLTK::Parser::Environment
+    class << self
+      def create(*args)
+        manifest = self.new
+        manifest.init(*args)
+        manifest
       end
     end
-  end
 
-  def payload
-    Hash[
-      @vars.map do |var, value|
-        [ var, value.respond_to?(:payload) ? value.payload : value ]
+    def init(token, project_name, manifest)
+      @token = token
+      @project_name = project_name
+      @manifest = manifest
+      @vars = {}
+    end
+
+    def macro(mac, args)
+      raise TypeError('Variable is not a macro') unless mac.is_a?(Archimedes::Macro)
+      resolve(mac.substitute(args))
+    end
+
+    def payload
+      fill_manifest
+
+      Hash[
+        @vars.map do |var, value|
+          [ var, value.respond_to?(:payload) ? value.payload : value ]
+        end
+      ]
+    end
+
+    private
+
+    def fill_manifest
+      @manifest.each do |variable, query|
+        begin
+          @vars[variable] = resolve(query)
+        rescue RLTK::NotInLanguage => e
+          raise LanguageError, "Could not resolve @#{variable} = #{query}"
+        end
       end
-    ]
-  end
+    end
 
-  private
-
-  def resolve(query)
-    InfixParser::parse(InfixLexer::lex(query), {env: @env})
+    def resolve(query)
+      Archimedes::InfixParser::parse(
+        Archimedes::InfixLexer::lex(query),
+        env: self
+      )
+    end
   end
 end
